@enhancement/86epe33fv/CS-SATA-Questionnaire-blocks ft adding return types
